make cpp files
