+++ include/varlink/service.hpp
@@ -4,7 +4,6 @@
 #include <mutex>
 #include <sstream>
 #include <varlink/detail/message.hpp>
-#include <varlink/detail/org.varlink.service.varlink.hpp>
 #include <varlink/interface.hpp>
 
 #define varlink_callback                               \
@@ -61,36 +60,7 @@ class varlink_service {
     }
 
   public:
-    explicit varlink_service(description Description) : desc(std::move(Description))
-    {
-        auto getInfo = [this] varlink_callback {
-            json::object_t info = {
-                {"vendor", desc.vendor},
-                {"product", desc.product},
-                {"version", desc.version},
-                {"url", desc.url}};
-            info["interfaces"] = json::array();
-            for (const auto& interface : interfaces) {
-                info["interfaces"].push_back(interface->name());
-            }
-            send_reply(info, false);
-        };
-        auto getInterfaceDescription = [this] varlink_callback {
-            const auto& ifname = parameters["interface"].get<std::string>();
-
-            if (const auto interface_it = find_interface(ifname); interface_it != interfaces.cend()) {
-                std::stringstream ss;
-                ss << **interface_it;
-                send_reply({{"description", ss.str()}}, false);
-            }
-            else {
-                throw varlink_error("org.varlink.service.InterfaceNotFound", {{"interface", ifname}});
-            }
-        };
-        add_interface(
-            org_varlink_service_varlink,
-            {{"GetInfo", getInfo}, {"GetInterfaceDescription", getInterfaceDescription}});
-    }
+    explicit varlink_service(description Description);
 
     varlink_service(const varlink_service& src) = delete;
     varlink_service& operator=(const varlink_service&) = delete;
